changed faulty dtype of wrapper arg

+++ wrapper.py
@@ -54,7 +54,7 @@ if __name__ == "__main__":
     parser.add_argument('--training-stage2-num-steps', type=int, default=1024,
                         help='lets you set the number of steps the sampler should take to sample images and generate the fid score')
 
-    parser.add_argument('--training-optimizer-weight-decay', type=int, default=0.015)
+    parser.add_argument('--training-optimizer-weight-decay', type=float, default=0.015)
 
     parser.add_argument('--training-verbosity', default='normal', choices=['silent', 'normal', 'verbose'],
                         help='lets you choose a verbosity mode for how much information will be logged.')
